code/requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.1
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bilibili-api-python==17.3.0
blis==1.3.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
cbor==1.0.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
cloudpathlib==0.21.1
cohere==5.16.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==4.0.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
duckdb==1.3.2
durationpy==0.10
effdet==0.4.1
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.11.0.post1
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
FlagEmbedding==1.3.5
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.8.0
grpcio==1.67.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.71
langchain-deepseek-official==0.1.0.post1
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.8
language_data==1.3.0
lark==1.2.2
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.51
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.52
llama-index-embeddings-adapter==0.3.0
llama-index-embeddings-huggingface==0.5.5
llama-index-embeddings-openai==0.3.1
llama-index-experimental==0.5.6
llama-index-finetuning==0.3.2
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.3.0
llama-index-llms-azure-openai==0.3.4
llama-index-llms-deepseek==0.1.2
llama-index-llms-mistralai==0.4.0
llama-index-llms-openai==0.4.7
llama-index-llms-openai-like==0.4.0
llama-index-multi-modal-llms-openai==0.5.3
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.6.43
lxml==5.4.0
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
milvus-lite==2.5.1
mistralai==1.9.2
mmh3==5.1.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.19.2
openai==1.97.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
peft==0.16.0
pi_heif==1.0.0
pikepdf==9.10.2
pillow==11.2.1
platformdirs==4.3.8
posthog==5.4.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycocotools==2.0.10
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.5.11
pymilvus.model==0.3.2
pyparsing==3.2.3
pypdf==5.8.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
qrcode-terminal==0.8
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.17
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.3
trec-car-tools==2.6
triton==3.3.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
unlzw3==0.2.3
unstructured==0.18.9
unstructured-client==0.39.1
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zlib-state==0.1.9
zstandard==0.23.0